NOISSUE: refactor level ledger tests
split internal and public API tests, with motivations:
1. protect from circular dependencies in future
2. test only public APIs
